+++ ChangeLog.md
@@ -281,6 +281,7 @@ Nazara Development Kit:
 - Added TextAreaWidget::[Get|Set]TextFont
 - ⚠️ TextAreaWidget::OnTextAreaCursorMove signal now uses a Vector2ui* position as its second argument (instead of a std::size_t*)
 - Added TextAreaWidget::OnTextAreaSelection
+- ⚠️ Console class is no longer bound to a LuaState and now has a OnCommand signal
 
 # 0.4: